[netlib] Reorganize the architecture
