+++ src/server/netlib/acceptor.h
@@ -10,17 +10,17 @@
 #ifndef MRPROXYER_SRC_NETLIB_ACCEPTOR_H_
 #define MRPROXYER_SRC_NETLIB_ACCEPTOR_H_
 
+#include "../util/non_copyable_movable.h"
 #include "io_event.h"
 #include "log.h"
 #include "net_address.h"
 #include "netlib.h"
-#include "non_copyable_movable.h"
 #include "socket.h"
 
 namespace netlib {
 class IoWatcher;
 
-class Acceptor : NonCopyableMovable {
+class Acceptor : util::NonCopyableMovable {
  public:
   Acceptor(IoWatcher* io_watcher, NetAddress bind_address);
   ~Acceptor();
