+++ src/server/netlib/poll.h
@@ -10,8 +10,8 @@
 #ifndef MRPROXYER_SRC_NETLIB_POLL_H_
 #define MRPROXYER_SRC_NETLIB_POLL_H_
 
+#include "../util/non_copyable_movable.h"
 #include "io_watcher.h"
-#include "non_copyable_movable.h"
 #include "platform.h"
 
 #ifdef OS_UNIX
